ClientInner refactor & TryFrom -> Deserialize in protocol
ClientInner now manages its connections internally, recreating them on
the fly with a factory function that the user provides.

A dedicated Deserialize trait was added since blanket impls can only be
done within the crate a trait is defined in.

The *Owned family of structs/traits was also reworked so higher rank
trait bounds work in `Client::receive_packet` now!